fix: resolve all CI lint failures
Backend:
- Extract issue watch methods to service_watch.go (reduces service.go
  from 861 to 657 effective lines, under revive file-length-limit 800)
- Extract OAuth credentials/helpers to service_oauth.go

Frontend:
- Format jira-page-client.tsx with prettier 3.8.1
