+++ apps/backend/internal/jira/service.go
@@ -12,9 +12,7 @@ import (
 
 	"github.com/kandev/kandev/internal/common/logger"
 	"github.com/kandev/kandev/internal/common/securityutil"
-	"github.com/kandev/kandev/internal/events"
 	"github.com/kandev/kandev/internal/events/bus"
-	"github.com/kandev/kandev/internal/task/repository/repoerrors"
 	"github.com/kandev/kandev/internal/watchreset"
 )
 
@@ -80,22 +78,6 @@ type Service struct {
 func (s *Service) SetWorkspaceAuthorizer(authorizer func(context.Context, string) error) {
 	if s != nil {
 		s.workspaceAuthorizer = authorizer
-	}
-}
-
-// RevealAccessToken reads the current OAuth access token from the secret store.
-// Implements the TokenRefresher interface for CloudClient auto-refresh.
-func (s *Service) RevealAccessToken(ctx context.Context, workspaceID string) (string, error) {
-	if s.secrets == nil {
-		return "", nil
-	}
-	return s.secrets.Reveal(ctx, OAuthAccessTokenKeyForWorkspace(workspaceID))
-}
-
-// SetOAuthRedirectURI sets the OAuth callback URL. Called at Provide time.
-func (s *Service) SetOAuthRedirectURI(uri string) {
-	if s != nil {
-		s.oauthRedirectURI = uri
 	}
 }
 
@@ -656,99 +638,6 @@ func (s *Service) invalidateClient(workspaceID string) {
 	s.mu.Unlock()
 }
 
-// resolveCredentials picks the credentials to test: if the request carries a
-// secret, use it inline (pre-save); otherwise fall back to the stored secret
-// (post-save re-test).
-func (s *Service) resolveCredentials(ctx context.Context, workspaceID string, req *SetConfigRequest) (*JiraConfig, string, error) {
-	cfg := &JiraConfig{
-		SiteURL:      normalizeSiteURL(req.SiteURL),
-		Email:        req.Email,
-		AuthMethod:   req.AuthMethod,
-		InstanceType: req.InstanceType,
-	}
-	var secret string
-	if req.Secret != "" {
-		secret = req.Secret
-	} else {
-		if s.secrets == nil {
-			return nil, "", errors.New("no secret store configured")
-		}
-		var err error
-		if cfg.AuthMethod == AuthMethodOAuth {
-			secret, err = s.secrets.Reveal(ctx, OAuthAccessTokenKeyForWorkspace(workspaceID))
-		} else {
-			secret, err = s.revealSecret(ctx, workspaceID)
-		}
-		if err != nil {
-			s.log.Warn("jira: secret reveal failed", zap.Error(err))
-			return nil, "", fmt.Errorf("read jira secret: %w", err)
-		}
-		if secret == "" {
-			return nil, "", errors.New("no token stored — paste one to test")
-		}
-	}
-	// Merge with persisted config only when the caller didn't pass enough on
-	// its own. A fully-specified pre-save TestConnection request needs no DB
-	// read — keeping the call gated avoids a transient DB error blocking auth
-	// testing for a request that doesn't depend on persisted state. When a
-	// read does happen, surface its error so a real DB failure isn't masked
-	// as an opaque "missing field" validation message later on.
-	needsStoredConfig := cfg.SiteURL == "" ||
-		cfg.AuthMethod == "" ||
-		cfg.InstanceType == "" ||
-		cfg.AuthMethod == AuthMethodOAuth ||
-		(cfg.AuthMethod == AuthMethodAPIToken && cfg.Email == "")
-	if needsStoredConfig {
-		if err := s.fillConfigFromStored(ctx, workspaceID, cfg); err != nil {
-			return nil, "", err
-		}
-	}
-	if cfg.InstanceType == "" {
-		cfg.InstanceType = InstanceTypeCloud
-	}
-	// Run the same auth/instance check the save path runs, so TestConnection
-	// reports invalid combos (e.g. pat+cloud, api_token+server) with a clear
-	// validation error instead of letting Jira return an opaque 401.
-	if err := validateAuthInstance(cfg.AuthMethod, cfg.InstanceType, cfg.Email); err != nil {
-		return nil, "", err
-	}
-	return cfg, secret, nil
-}
-
-// fillConfigFromStored reads the persisted singleton config and copies any
-// fields the caller left blank onto cfg. A real read failure is returned so
-// callers don't mask DB errors as opaque "missing field" validation errors;
-// a missing row (stored == nil) is treated as "nothing to fill" rather than
-// an error so first-time TestConnection requests still go through.
-func (s *Service) fillConfigFromStored(ctx context.Context, workspaceID string, cfg *JiraConfig) error {
-	stored, err := s.store.GetConfigForWorkspace(ctx, workspaceID)
-	if err != nil {
-		return fmt.Errorf("read jira config: %w", err)
-	}
-	if stored == nil {
-		return nil
-	}
-	if cfg.SiteURL == "" {
-		cfg.SiteURL = stored.SiteURL
-	}
-	if cfg.Email == "" {
-		cfg.Email = stored.Email
-	}
-	if cfg.AuthMethod == "" {
-		cfg.AuthMethod = stored.AuthMethod
-	}
-	if cfg.InstanceType == "" {
-		cfg.InstanceType = stored.InstanceType
-	}
-	if cfg.ClientID == "" {
-		cfg.ClientID = stored.ClientID
-	}
-	if cfg.CloudID == "" {
-		cfg.CloudID = stored.CloudID
-	}
-	return nil
-}
-
 func (s *Service) defaultWorkspaceID() (string, error) {
 	return s.store.defaultWorkspaceID()
 }
@@ -897,286 +786,6 @@ func (s *Service) SetEventBus(eb bus.EventBus) {
 
 // ErrIssueWatchNotFound is returned when GetIssueWatch's caller looks up an ID
 // that doesn't exist. Callers map this to HTTP 404.
-var ErrIssueWatchNotFound = errors.New("jira: issue watch not found")
-
-// CreateIssueWatch validates the request and persists a new watch row.
-func (s *Service) CreateIssueWatch(ctx context.Context, req *CreateIssueWatchRequest) (*IssueWatch, error) {
-	if err := validateIssueWatchCreate(req); err != nil {
-		return nil, err
-	}
-	// WorkspaceID comes from the request body, which the query-only integration
-	// middleware never authorizes. Without this a caller could plant a watch in
-	// another user's workspace that repeatedly spawns tasks with that user's
-	// credentials and an attacker-supplied JQL/prompt.
-	if err := s.authorizeWorkspaceAccess(ctx, req.WorkspaceID); err != nil {
-		return nil, err
-	}
-	repositoryID, baseBranch, err := s.resolveRepositoryBinding(ctx, req.WorkspaceID, req.RepositoryID, req.BaseBranch)
-	if err != nil {
-		return nil, err
-	}
-	w := &IssueWatch{
-		WorkspaceID:         req.WorkspaceID,
-		WorkflowID:          req.WorkflowID,
-		WorkflowStepID:      req.WorkflowStepID,
-		RepositoryID:        repositoryID,
-		BaseBranch:          baseBranch,
-		JQL:                 strings.TrimSpace(req.JQL),
-		AgentProfileID:      req.AgentProfileID,
-		ExecutorProfileID:   req.ExecutorProfileID,
-		Prompt:              req.Prompt,
-		PollIntervalSeconds: req.PollIntervalSeconds,
-		MaxInflightTasks:    req.MaxInflightTasks,
-		Enabled:             true,
-	}
-	if req.Enabled != nil {
-		w.Enabled = *req.Enabled
-	}
-	if err := s.store.CreateIssueWatch(ctx, w); err != nil {
-		return nil, err
-	}
-	return w, nil
-}
-
-// ListIssueWatches returns the watches configured for a workspace.
-func (s *Service) ListIssueWatches(ctx context.Context, workspaceID string) ([]*IssueWatch, error) {
-	if err := s.authorizeWorkspaceAccess(ctx, workspaceID); err != nil {
-		return nil, err
-	}
-	return s.store.ListIssueWatches(ctx, workspaceID)
-}
-
-// ListAllIssueWatches returns every watch the caller may see. For a scoped
-// caller that is only their own workspaces' watches; for an identity-less
-// internal caller (unscoped) it is every watch, as before auth. The unscoped
-// list form is only reachable when workspace_id is omitted, which the query-only
-// integration middleware does not authorize — without this filter it leaked
-// every workspace's watch config (JQL, repo/agent/profile IDs, spawn prompt).
-func (s *Service) ListAllIssueWatches(ctx context.Context) ([]*IssueWatch, error) {
-	watches, err := s.store.ListAllIssueWatches(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return s.filterIssueWatchesByAccess(ctx, watches)
-}
-
-// filterIssueWatchesByAccess keeps only watches whose workspace the caller may
-// access. Access decisions are memoized per workspace so a long list costs one
-// authorize call per distinct workspace, not one per watch. Only an
-// ErrWorkspaceNotFound denial drops a watch; any other authorizer error (a
-// transient DB failure, say) is propagated so the caller sees the failure
-// rather than a silently truncated 200.
-func (s *Service) filterIssueWatchesByAccess(ctx context.Context, watches []*IssueWatch) ([]*IssueWatch, error) {
-	decision := make(map[string]bool)
-	visible := make([]*IssueWatch, 0, len(watches))
-	for _, w := range watches {
-		allowed, seen := decision[w.WorkspaceID]
-		if !seen {
-			switch err := s.authorizeWorkspaceAccess(ctx, w.WorkspaceID); {
-			case err == nil:
-				allowed = true
-			case errors.Is(err, repoerrors.ErrWorkspaceNotFound):
-				allowed = false
-			default:
-				return nil, err
-			}
-			decision[w.WorkspaceID] = allowed
-		}
-		if allowed {
-			visible = append(visible, w)
-		}
-	}
-	return visible, nil
-}
-
-// GetIssueWatch returns a single watch by ID or ErrIssueWatchNotFound.
-func (s *Service) GetIssueWatch(ctx context.Context, id string) (*IssueWatch, error) {
-	w, err := s.store.GetIssueWatch(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-	if w == nil {
-		return nil, ErrIssueWatchNotFound
-	}
-	return w, nil
-}
-
-// UpdateIssueWatch applies a partial update by patching only the fields the
-// caller explicitly set, then persists the result.
-func (s *Service) UpdateIssueWatch(ctx context.Context, id string, req *UpdateIssueWatchRequest) (*IssueWatch, error) {
-	w, err := s.GetIssueWatch(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-	// Authorize off the loaded row's workspace — the watch ID alone reveals
-	// nothing about ownership, so a caller must not mutate another user's watch.
-	if err := s.authorizeWorkspaceAccess(ctx, w.WorkspaceID); err != nil {
-		return nil, err
-	}
-	prevRepositoryID, prevBaseBranch := w.RepositoryID, w.BaseBranch
-	applyIssueWatchPatch(w, req)
-	// Empty-string PATCH writes (`{"workflowId": ""}` etc.) bypass the nil
-	// guard in applyIssueWatchPatch — Go's JSON decoder returns a non-nil
-	// pointer to "". Guard the post-patch state so a PATCH can't strip a row
-	// of the fields the orchestrator needs to create tasks.
-	if w.JQL == "" {
-		return nil, fmt.Errorf("%w: jql cannot be empty", ErrInvalidConfig)
-	}
-	if w.WorkflowID == "" || w.WorkflowStepID == "" {
-		return nil, fmt.Errorf("%w: workflowId and workflowStepId cannot be empty", ErrInvalidConfig)
-	}
-	if err := validatePollInterval(w.PollIntervalSeconds); err != nil {
-		return nil, err
-	}
-	if err := validateMaxInflightTasks(w.MaxInflightTasks); err != nil {
-		return nil, err
-	}
-	// Only validate/resolve the binding when its value actually changed. The
-	// dialog re-sends repositoryId/baseBranch on every PATCH, and an unchanged
-	// binding whose repo was since soft-deleted must not block edits to other
-	// fields (prompt, JQL, …).
-	if w.RepositoryID != prevRepositoryID || w.BaseBranch != prevBaseBranch {
-		repositoryID, baseBranch, err := s.resolveRepositoryBinding(ctx, w.WorkspaceID, w.RepositoryID, w.BaseBranch)
-		if err != nil {
-			return nil, err
-		}
-		w.RepositoryID = repositoryID
-		w.BaseBranch = baseBranch
-	}
-	if err := s.store.UpdateIssueWatch(ctx, w); err != nil {
-		return nil, err
-	}
-	return w, nil
-}
-
-// DeleteIssueWatch removes the watch and its dedup rows. Idempotent: deleting
-// a missing ID is a silent success.
-func (s *Service) DeleteIssueWatch(ctx context.Context, id string) error {
-	return s.store.DeleteIssueWatch(ctx, id)
-}
-
-// jiraIssueWatchResetter is the watchreset.Resetter adapter for a single
-// Jira issue watch. Closes over the store and watch ID so the shared
-// watchreset.Run helper stays integration-agnostic.
-type jiraIssueWatchResetter struct {
-	store   *Store
-	watchID string
-}
-
-func (r *jiraIssueWatchResetter) ListTaskIDs(ctx context.Context) ([]string, error) {
-	return r.store.ListIssueWatchTaskIDs(ctx, r.watchID)
-}
-
-func (r *jiraIssueWatchResetter) Clear(ctx context.Context) error {
-	return r.store.ResetIssueWatchState(ctx, r.watchID)
-}
-
-// PreviewResetIssueWatch returns how many tasks ResetIssueWatch would
-// cascade-delete. Used by the frontend to populate the confirmation dialog.
-func (s *Service) PreviewResetIssueWatch(ctx context.Context, watchID string) (int, error) {
-	return watchreset.Preview(ctx, &jiraIssueWatchResetter{store: s.store, watchID: watchID})
-}
-
-// ResetIssueWatch is destructive: cascade-deletes every task previously
-// created by the watch (including archived), wipes the per-watch dedup
-// rows, and nulls last_polled_at so the next poll re-imports every
-// currently-matching ticket. Returns the count of tasks deleted.
-func (s *Service) ResetIssueWatch(ctx context.Context, watchID string) (int, error) {
-	s.mu.Lock()
-	td := s.taskDeleter
-	s.mu.Unlock()
-	if td == nil {
-		return 0, errors.New("jira: task deleter not wired; reset unavailable")
-	}
-	res, err := watchreset.Run(ctx,
-		&jiraIssueWatchResetter{store: s.store, watchID: watchID},
-		td, s.log)
-	return res.TasksDeleted, err
-}
-
-// CheckIssueWatch runs the watch's JQL once and returns the tickets that
-// haven't been turned into tasks yet. last_polled_at is stamped regardless of
-// whether the search succeeded — a failing search still counts as "we tried"
-// so the UI can show liveness.
-func (s *Service) CheckIssueWatch(ctx context.Context, w *IssueWatch) ([]*JiraTicket, error) {
-	defer s.stampLastPolled(w.ID)
-	client, err := s.clientFor(ctx, w.WorkspaceID)
-	if err != nil {
-		return nil, err
-	}
-	// Single page is enough per tick: the next poll picks up anything that
-	// overflowed maxResults. Pagination would let one chatty workspace starve
-	// the others without any real freshness benefit.
-	res, err := client.SearchTickets(ctx, w.JQL, "", issueWatchSearchPageSize)
-	if err != nil {
-		return nil, err
-	}
-	// Bulk-fetch the dedup set once per call instead of one query per ticket —
-	// a broad JQL can return up to issueWatchSearchPageSize tickets, so a
-	// per-ticket round-trip would multiply the per-tick cost.
-	seen, err := s.store.ListSeenIssueKeys(ctx, w.ID)
-	if err != nil {
-		s.log.Warn("jira: dedup set fetch failed",
-			zap.String("watch_id", w.ID), zap.Error(err))
-		seen = nil // fall through: a missing dedup set is safer than dropping all tickets
-	}
-	out := make([]*JiraTicket, 0, len(res.Tickets))
-	for i := range res.Tickets {
-		t := res.Tickets[i]
-		if _, ok := seen[t.Key]; ok {
-			continue
-		}
-		out = append(out, &t)
-	}
-	return out, nil
-}
-
-// stampLastPolled writes the current timestamp onto the watch row using a
-// fresh background context with a short write deadline. The caller's ctx may
-// already be cancelled (typical at poller shutdown), which would cause the
-// DB write to fail silently and the watch would look "never polled" on the
-// next backend restart. Mirrors the pattern in RecordAuthHealth.
-func (s *Service) stampLastPolled(watchID string) {
-	ctx, cancel := context.WithTimeout(context.Background(), authHealthWriteTimeout)
-	defer cancel()
-	if err := s.store.UpdateIssueWatchLastPolled(ctx, watchID, time.Now().UTC()); err != nil {
-		s.log.Warn("jira: update last_polled_at failed",
-			zap.String("watch_id", watchID), zap.Error(err))
-	}
-}
-
-// publishNewJiraIssueEvent emits the orchestrator-facing event for one freshly
-// observed ticket. No-op when the event bus is not wired (tests, early boot).
-func (s *Service) publishNewJiraIssueEvent(ctx context.Context, w *IssueWatch, ticket *JiraTicket) {
-	s.mu.Lock()
-	eb := s.eventBus
-	s.mu.Unlock()
-	if eb == nil {
-		return
-	}
-	evt := bus.NewEvent(events.JiraNewIssue, "jira", &NewJiraIssueEvent{
-		IssueWatchID:      w.ID,
-		WorkspaceID:       w.WorkspaceID,
-		WorkflowID:        w.WorkflowID,
-		WorkflowStepID:    w.WorkflowStepID,
-		RepositoryID:      w.RepositoryID,
-		BaseBranch:        w.BaseBranch,
-		AgentProfileID:    w.AgentProfileID,
-		ExecutorProfileID: w.ExecutorProfileID,
-		Prompt:            w.Prompt,
-		MaxInflightTasks:  w.MaxInflightTasks,
-		Issue:             ticket,
-	})
-	if err := eb.Publish(ctx, events.JiraNewIssue, evt); err != nil {
-		s.log.Debug("jira: publish new issue event failed",
-			zap.String("watch_id", w.ID), zap.String("issue_key", ticket.Key), zap.Error(err))
-	}
-}
-
-// issueWatchSearchPageSize caps how many tickets a single CheckIssueWatch call
-// pulls from JIRA. 50 is well under the API's 100-result limit and keeps the
-// per-tick cost bounded even for very broad JQL queries.
-const issueWatchSearchPageSize = 50
 
 // MinIssueWatchPollInterval / MaxIssueWatchPollInterval bound the per-watch
 // JQL re-run cadence. The lower limit protects Atlassian from rapid-fire
@@ -1311,4 +920,3 @@ func applyIssueWatchPatch(w *IssueWatch, req *UpdateIssueWatchRequest) {
 		w.MaxInflightTasks = req.MaxInflightTasks.Value
 	}
 }
-
