+++ apps/web/app/jira/jira-page-client.tsx
@@ -385,4 +385,3 @@ export function JiraPageClient({ workspaceId, workflows, steps }: JiraPageClient
     </PageShell>
   );
 }
-
